refactor(frontend): organize imports to use specific paths instead of barrel exports
Update import statements across the frontend to import from specific module
paths rather than using barrel exports from ~/components/common. This improves
tree-shaking and makes dependencies more explicit.

Changes:
- Remove barrel exports from ~/components/common/index.tsx
- Update imports across routes and components to use specific paths
- Import from entity-display, filters, layout, review, summary modules directly
