+++ apps/frontend/app/components/common/index.tsx
@@ -500,10 +500,3 @@ export const CollectionTemplateRenderer = ({
 		</Fragment>
 	);
 };
-
-export * from "./entity-display";
-export * from "./filters";
-export * from "./layout";
-export * from "./multi-select-creatable";
-export * from "./review";
-export * from "./summary";
