+++ apps/frontend/app/routes/_dashboard.media.genre.$id._index.tsx
@@ -15,7 +15,7 @@ import {
 } from "@ryot/ts-utils";
 import { useLoaderData } from "react-router";
 import { z } from "zod";
-import { ApplicationGrid } from "~/components/common";
+import { ApplicationGrid } from "~/components/common/layout";
 import { MetadataDisplayItem } from "~/components/media/display-items";
 import { pageQueryParam } from "~/lib/shared/constants";
 import { useAppSearchParam } from "~/lib/shared/hooks";
