+++ apps/frontend/app/routes/_dashboard.media.people.$action.tsx
@@ -47,13 +47,15 @@ import { $path } from "safe-routes";
 import { match } from "ts-pattern";
 import { z } from "zod";
 import {
-	ApplicationGrid,
 	BulkCollectionEditingAffix,
+	DisplayListDetailsAndRefresh,
+} from "~/components/common";
+import {
 	CollectionsFilter,
 	DebouncedSearchInput,
-	DisplayListDetailsAndRefresh,
 	FiltersModal,
-} from "~/components/common";
+} from "~/components/common/filters";
+import { ApplicationGrid } from "~/components/common/layout";
 import { PersonDisplayItem } from "~/components/media/display-items";
 import { pageQueryParam } from "~/lib/shared/constants";
 import { useAppSearchParam, useCoreDetails } from "~/lib/shared/hooks";
