+++ apps/frontend/app/routes/_dashboard.settings.users.tsx
@@ -59,7 +59,7 @@ import { $path } from "safe-routes";
 import { match } from "ts-pattern";
 import { withQuery } from "ufo";
 import { z } from "zod";
-import { DebouncedSearchInput } from "~/components/common";
+import { DebouncedSearchInput } from "~/components/common/filters";
 import { useCoreDetails } from "~/lib/shared/hooks";
 import { clientGqlService } from "~/lib/shared/query-factory";
 import { openConfirmationModal } from "~/lib/shared/ui-utils";
